requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
flightsql-dbapi==0.2.1
idna==3.4
numpy==1.25.0
pandas==2.0.2
protobuf==4.23.3
pyarrow==12.0.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
SQLAlchemy==1.4.48
tzdata==2023.3
urllib3==2.0.3
